Add files via upload
Scoreboard fully functional 

Off-by-one-error:
Scenario/Problem: Each score would update except last (fifth) place, that would never change.
Solution:This was because I made an array that held the updated five high scores and five high scorers. Since the updated list does not contain the now-sixth place high score, I wasn't able to use it to tell the file what to look for to overwrite it. I now have separate arrays for "previous high scores" and "updated high scores" so that I don't delete any information too soon, solving the issue

Overwriting the scores:
Scenario/Problem - if a new high score came in and "15" had to replace the position below it where "10" used to be, my code told it to find the first instance of "10" and overwrite it with "15". The problem was that scores such as "100" would be overwritten as "150" since it includes a 10. 
Solution: Since all of the scores are surrounded in bold tags, I searched the document for "<b>10</b>" so that it would skip over larger scores that happened to contain the score we wanted to overwrite.